Rename common/controller-factories to common/controllers
